requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
altgraph==0.17.4
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
flake8==7.3.0
frozenlist==1.8.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
mccabe==0.7.0
moviepy==2.2.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
proglog==0.1.12
propcache==0.4.1
pushbullet.py==0.12.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
python-dotenv==1.1.1
python-magic==0.4.27
python-telegram-bot==22.5
pytokens==0.1.10
qasync==0.28.0
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
tqdm==4.67.1
twilio==9.8.3
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
yarl==1.22.0
